requirements.txt
torch
torchaudio
lmdb
einops_exts
torch_ema
nnAudio
tensorboard
numpy>=2.0
gin-config
pretty_midi
tqdm>=4.64.1
cached-conv>=2.5.0
cqt_pytorch==0.0.5
matplotlib>=3.8.3
librosa>=0.10.2
pyyaml
demucs